+++ skcuda/misc.py
@@ -368,7 +368,7 @@ def select_block_grid_sizes(dev, data_shape, threads_per_block=None):
     else:
         raise ValueError('array size too large')
 
-def zeros(shape, dtype, allocator=drv.mem_alloc):
+def zeros(shape, dtype, order='C', allocator=drv.mem_alloc):
     """
     Return an array of the given shape and dtype filled with zeros.
 
@@ -378,14 +378,16 @@ def zeros(shape, dtype, allocator=drv.mem_alloc):
         Array shape.
     dtype : data-type
         Data type for the array.
-    allocator : callable
+    order : {'C', 'F'}, optional
+        Create array using row-major or column-major format.
+    allocator : callable, optional
         Returns an object that represents the memory allocated for
         the requested array.
 
     Returns
     -------
     out : pycuda.gpuarray.GPUArray
-        Array of zeros with the given shape and dtype.
+        Array of zeros with the given shape, dtype, and order.
 
     Notes
     -----
@@ -395,7 +397,7 @@ def zeros(shape, dtype, allocator=drv.mem_alloc):
     http://projects.scipy.org/numpy/ticket/1898
     """
 
-    out = gpuarray.GPUArray(shape, dtype, allocator)
+    out = gpuarray.GPUArray(shape, dtype, allocator, order=order)
     out.fill(0)
     return out
 
@@ -413,14 +415,15 @@ def zeros_like(a):
     Returns
     -------
     out : pycuda.gpuarray.GPUArray
-        Array of zeros with the shape and dtype of `a`.
+        Array of zeros with the shape, dtype, and strides of `a`.
     """
 
-    out = gpuarray.GPUArray(a.shape, a.dtype, drv.mem_alloc)
+    out = gpuarray.GPUArray(a.shape, a.dtype, drv.mem_alloc,
+                            strides=a.strides)
     out.fill(0)
     return out
 
-def ones(shape, dtype, allocator=drv.mem_alloc):
+def ones(shape, dtype, order='C', allocator=drv.mem_alloc):
     """
     Return an array of the given shape and dtype filled with ones.
 
@@ -430,41 +433,44 @@ def ones(shape, dtype, allocator=drv.mem_alloc):
         Array shape.
     dtype : data-type
         Data type for the array.
-    allocator : callable
+    order : {'C', 'F'}, optional
+        Create array using row-major or column-major format.
+    allocator : callable, optional
         Returns an object that represents the memory allocated for
         the requested array.
 
     Returns
     -------
     out : pycuda.gpuarray.GPUArray
-        Array of ones with the given shape and dtype.
+        Array of ones with the given shape, dtype, and order.
     """
 
-    out = gpuarray.GPUArray(shape, dtype, allocator)
+    out = gpuarray.GPUArray(shape, dtype, allocator, order=order)
     out.fill(1)
     return out
 
-def ones_like(other):
+def ones_like(a):
     """
     Return an array of ones with the same shape and type as a given array.
 
     Parameters
     ----------
-    other : pycuda.gpuarray.GPUArray
-        Array whose shape and dtype are to be used to allocate a new array.
+    a : array_like
+        The shape and data type of `a` determine the corresponding
+        attributes of the returned array.
 
     Returns
     -------
     out : pycuda.gpuarray.GPUArray
-        Array of ones with the shape and dtype of `other`.
+        Array of ones with the shape, dtype, and strides of `other`.
     """
 
-    out = gpuarray.GPUArray(other.shape, other.dtype,
-                            other.allocator)
+    out = gpuarray.GPUArray(a.shape, a.dtype,
+                            a.allocator, strides=a.strides)
     out.fill(1)
     return out
 
-def inf(shape, dtype, allocator=drv.mem_alloc):
+def inf(shape, dtype, order='C', allocator=drv.mem_alloc):
     """
     Return an array of the given shape and dtype filled with infs.
 
@@ -474,17 +480,19 @@ def inf(shape, dtype, allocator=drv.mem_alloc):
         Array shape.
     dtype : data-type
         Data type for the array.
-    allocator : callable
+    order : {'C', 'F'}, optional
+        Create array using row-major or column-major format.
+    allocator : callable, optional
         Returns an object that represents the memory allocated for
         the requested array.
 
     Returns
     -------
     out : pycuda.gpuarray.GPUArray
-        Array of infs with the given shape and dtype.
+        Array of infs with the given shape, dtype, and order.
     """
 
-    out = gpuarray.GPUArray(shape, dtype, allocator)
+    out = gpuarray.GPUArray(shape, dtype, allocator, order=order)
     out.fill(np.inf)
     return out